+++ test/integration/tests/checkquote.sh
@@ -59,4 +59,18 @@ tpm2_quote -c $handle_ak -l sha256:15,16,22 -q $loaded_randomness \
 tpm2_checkquote -u $output_ak_pub_pem -m $output_quote -s $output_quotesig \
 -f $output_quotepcr -g $digestAlg -q $loaded_randomness
 
+# Verify EC
+
+tpm2_createek -G ecc -c ecc.ek
+
+tpm2_createak -C ecc.ek -c ecc.ak -G ecc -g sha256 -s ecdsa
+
+tpm2_readpublic -c ecc.ak -f pem -o ecc.ek.pem
+
+tpm2_getrandom -o nonce.bin 20
+
+tpm2_quote -c ecc.ak -l sha256:15,16,22 -q nonce.bin -m quote.bin -s quote.sig -o quote.pcr -g sha256
+
+tpm2_checkquote -u ecc.ek.pem -m quote.bin -s quote.sig -f quote.pcr -g sha256 -q nonce.bin
+
 exit 0